Add 2332 MultiROM PLCC
This has the same functionality as the other switcher but uses a Winbond PLCC EEPROM which are available, inexpensive, and smaller in size than the DIP version. The pcb is the same width as the DIP socket and only slightly longer meaning it will interfere less with other surrounding parts, and will also work in the 8296 motherboards so you can install two side-by-side in the OptionROM sockets, plus they will not interfere with the expansion headers.